refactor(agent_program/goal_curator): split tests to separate file (#1266)
Splits src/agent_program/goal_curator.rs (408 LOC) into:
- src/agent_program/goal_curator/mod.rs   (~180 LOC) — production
- src/agent_program/goal_curator/tests.rs (~228 LOC) — test module

Pure refactor: zero logic changes, all 20 goal_curator tests pass.

Closes 1/70 of #1266 (after #1285-#1288).
